setopt: split out curl_easy_setopt() to its own file
... to make url.c smaller.
